Move enum parsers to subdirectory
