+++ source/dotnet/Measurements.Application/Responses/GetAggregatedMeasurementsResponse.cs
@@ -3,6 +3,7 @@ using System.Text.Json.Serialization;
 using Energinet.DataHub.Measurements.Application.Exceptions;
 using Energinet.DataHub.Measurements.Application.Extensions;
 using Energinet.DataHub.Measurements.Application.Persistence;
+using Energinet.DataHub.Measurements.Application.Responses.EnumParsers;
 using Energinet.DataHub.Measurements.Domain;
 
 namespace Energinet.DataHub.Measurements.Application.Responses;
