+++ source/dotnet/Measurements.UnitTests/Application/Responses/ResolutionParserTests.cs
@@ -1,4 +1,5 @@
 ﻿using Energinet.DataHub.Measurements.Application.Responses;
+using Energinet.DataHub.Measurements.Application.Responses.EnumParsers;
 using Energinet.DataHub.Measurements.Domain;
 using Xunit;
 using Xunit.Categories;
